requirements.dev.txt
# dev dependencies
ipython
ipdb
pudb
pytest-watch

# black --check --diff dundie tests integrationCode quality
flake8
pyproject-flake8
black

# install project as editable
-e .
